Refactor ray tracing and shadowing modules; remove legacy code and implement new structure
- Deleted old ray and shadowing implementations in `src/rays.py` and `src/shadowing.py`.
- Introduced new ray and shadowing logic in `python/modules/rays.py` and `python/modules/shadowing.py`.
- Updated utility functions in `python/modules/util.py` for better modularity.
- Reorganized variable definitions in `python/modules/variables.py` for clarity.
- Added new object classes for emitters and absorbers in `python/modules/objects.py`.
- Implemented a basic game loop in `python/main.py` to handle user input and rendering.
- Integrated `macroquad` for rendering in Rust with a simple example in `src/main.rs`.
- Updated dependencies in `Cargo.toml` and `python/requirements.txt` to reflect new libraries.
