+++ .gitignore
@@ -5,3 +5,14 @@ __pycache__/
 
 # vscode
 .vscode/
+
+# rust
+# Generated by Cargo
+# will have compiled files and executables
+debug/
+target/
+
+# These are backup files generated by rustfmt
+**/*.rs.bk
+
+*.pdb
