+++ Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "raytracer"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+macroquad = "0.4.14"
